requirements.txt
allure-pytest==2.15.3
flake8==7.3.0
pytest==9.0.3
requests==2.33.1
selenium==4.43.0
tenacity==9.1.4
webdriver-manager==4.0.2
